+++ include/odroid/digital/digital.hpp
@@ -5,8 +5,8 @@
 #include <memory>
 
 namespace gpio::digital {
-enum class Mode : uint8_t { INPUT = 0, OUTPUT = 1, OFF = 0 };
-enum class Write : uint8_t { LOW = 0, HIGH = 1 };
+enum class Mode : uint16_t { INPUT = 0, OUTPUT = 1, OFF = 0 };
+enum class Write : uint16_t { LOW = 0, HIGH = 1 };
 } // namespace gpio::digital
 
 #endif // ODROIDGPIO_DIGITAL_HPP
