+++ include/odroid/pwm/pwm.hpp
@@ -6,7 +6,7 @@
 namespace gpio::pwm {
 enum class Mode : int { OUTPUT = 4, SOFT_OUTPUT = 6, TONE_OUTPUTOFF = 8, OFF = 0 };
 
-void clock(uint8_t hz);
+void clock(uint16_t hz);
 void range(uint16_t range);
 } // namespace gpio::pwm
