+++ src/digital/PWMPin.cpp
@@ -3,15 +3,15 @@
 #include <odroid/digital/DigitalPin.hpp>
 #include <wiringPi.h>
 
-gpio::digital::DigitalPin::DigitalPin(uint8_t pin_number, Mode mode) :
+gpio::digital::DigitalPin::DigitalPin(uint16_t pin_number, Mode mode) :
    BasePin(pin_number), m_mode(mode)
 {
-   gpio::pin_mode(m_pin_number, static_cast<uint8_t>(m_mode));
+   gpio::pin_mode(m_pin_number, static_cast<uint16_t>(m_mode));
 }
 
 gpio::digital::DigitalPin::~DigitalPin()
 {
-   gpio::pin_mode(m_pin_number, static_cast<uint8_t>(Mode::OFF));
+   gpio::pin_mode(m_pin_number, static_cast<uint16_t>(Mode::OFF));
 }
 
 auto gpio::digital::DigitalPin::mode() const -> Mode
@@ -22,7 +22,7 @@ auto gpio::digital::DigitalPin::mode() const -> Mode
 void gpio::digital::DigitalPin::mode(Mode mode)
 {
    m_mode = mode;
-   gpio::pin_mode(m_pin_number, static_cast<uint8_t>(m_mode));
+   gpio::pin_mode(m_pin_number, static_cast<uint16_t>(m_mode));
 }
 
 void gpio::digital::DigitalPin::write(Write mode)
