+++ src/digital/digital.cpp
@@ -2,7 +2,7 @@
 
 #include <wiringPi.h>
 
-void gpio::digital::write(uint8_t pin_number, Write mode)
+void gpio::digital::write(uint16_t pin_number, Write mode)
 {
    digitalWrite(pin_number, static_cast<int>(mode));
 }
