+++ src/include/detail/gpio.hpp
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 namespace gpio {
-void pin_mode(uint8_t pin_number, uint8_t mode);
+void pin_mode(uint16_t pin_number, uint16_t mode);
 } // namespace gpio
 
 #endif // DETAIL_ODROIDGPIO_GPIO_HPP
